Shorthand for picking speeds and then recalculating object

Picks unreal speeds in the same way as pick_unreal_speed` but then removes given lines form the data, replaces them as per replacement parameter and returnes cleaned object

remove_unreal_speeds(obj, ...)

# S3 method for navr
remove_unreal_speeds(obj, cutoff = NULL, type = NULL,
  remove_distance = TRUE, total_recalculate = TRUE, indices = c(),
  replacement = NA)

Arguments

obj
valid navr object. See remvoe_unreal_speeds.navr

...

cutoff
cutoff value. see pick_unreal_speeds

type
what type of cutoof. see pick_unreal_speeds

remove_distance
shoudl the distances be removed? in case unreal speeds are probably due to "teleportation" or bad measurements, jumpy tracking etc, then the participant didn't really travel that distance and we can remove it. In case the unreal distances are caused by parts of the tracking missing, then the distances travelled are actually correct, just the speeds are not. Defaults to TRUE

total_recalculate
if true, recalculates total_distance column to reflect removed distances, Defaults to TRUE

indices
indices of speeds to clean out in case they were obtained separately. If empty, type, cutoff need to be defined and pick_unreal_speeds is called

replacement
what to replace unreal speeds with. Defaults to NA

Methods (by class)

navr: clears out navr object